Changes in boxfish::download_manager API and introduction of boxfish::download_source
Changed boxfish::download_manager API to have begin_open() that must be called
on UI thread (one that owns QNetworkAccessManager) and end_open() that must be
called in a separate thread or upon receival opened() signal.

Created a new class boxfish::download_source which is a simple adapter for
using the download_manager with boost::iostreams.